Move atomic functions from constants to atomic
Up until now, atomic functions were in the constants subpackage, which
isn't really an intuitive place for them.  This commit moves them into
a new atomic subpackage.
