+++ plasmapy/atomic/__init__.py
@@ -1,0 +1,11 @@
+"""Physical and mathematical constants for use within PlasmaPy"""
+from .atomic import (element_symbol, isotope_symbol, atomic_number,
+                     mass_number, element_name, standard_atomic_weight,
+                     isotope_mass, ion_mass, nuclear_binding_energy, half_life,
+                     energy_from_nuclear_reaction, is_isotope_stable,
+                     known_isotopes, common_isotopes, stable_isotopes,
+                     isotopic_abundance, charge_state)
+
+from .elements import Elements
+
+from .isotopes import Isotopes
