+++ plasmapy/atomic/isotopes.py
