+++ plasmapy/constants/__init__.py
@@ -8,16 +8,3 @@ from astropy.constants.si import (h, hbar, k_B, c, G, g0, m_p, m_n, m_e,
                                   M_sun, R_sun, M_earth, R_earth)
 
 from astropy.constants import atm
-
-from .atomic import (element_symbol, isotope_symbol, atomic_number,
-                     mass_number, element_name, standard_atomic_weight,
-                     isotope_mass, ion_mass, nuclear_binding_energy, half_life,
-                     energy_from_nuclear_reaction, is_isotope_stable,
-                     known_isotopes, common_isotopes, stable_isotopes,
-                     isotopic_abundance, charge_state)
-
-from .elements import Elements
-
-from .isotopes import Isotopes
-
-from . import tests
